Create SocialRankingSolution

Map a power relation between coalitions to a power relation between elements, also known as a social ranking solution.

Usage

doRanking(scores, compare = NULL, decreasing = TRUE)

Arguments

scores

A sortable vector or list of element scores. If names(scores) is not NULL, those will be used as element names. Else a number sequence corresponding to the elements is generated.

compare

Optional comparison function taking in two elements and returning a numerical value based on the relation between these two elements. If set to NULL, the default order() function is called. See details for more information.

decreasing

If TRUE (default), elements with higher scores are ranked higher.

Value

A list of type SocialRankingSolution. Each element of the list contains a sets::set() of elements in powerRelation that are indifferent to one another.

Details

A normal sort function is used to get the order of scores and determine the ranking from highest to lowest scoring element (or lowest to highest scoring if decreasing is set to FALSE). This works in instances where the score vector contains simple numerical values that are easily sortable.

In more complex scenarios it may be necessary to introduce a custom comparison function to determine the relation between elements. In the case of LexcelScore, a custom S3 class is introduced that implements the [, == and > operator (see examples).

To circumvent the introduction of S3 classes, a function can be assigned to the compare parameter. It must take two parameters (i.e. function(a, b)) and returns one of the following:

> 0 (positive value) if score a is ranked higher than score b

< 0 (negative value) if score a is ranked lower than score b

= 0 if both scores a and b are considered equal.

This means for instance that doRanking(c(a=3,b=2,c=2), compare = function(a,b) a - b) ranks elements with higher scores higher (here a > b ~ c), whereas doRanking(c(a=3,b=2,c=2), compare = function(a,b) b - a) favors elements with lower scores (b ~ c > a).

If the compare is set to NULL, the default order() function is called. This means that for lists an S3 class has to be implemented supporting the indexing and comparison operators. See examples for more.

Examples

# TODO work on examples
pr <- as.PowerRelation("2 > 12 > 1")

# we define our own social ranking solution.
# a player's score is determined by the equivalence class index it first appears in.
# lower is better
scores <- c(`1` = 2, `2` = 1)

# 2 > 1
doRanking(scores)
#> 1 > 2