SocialRanking object

Create a SocialRanking object.

Usage

SocialRanking(l)

Arguments

l

A list of vectors

Value

A list of type SocialRanking. Each element of the list contains a vector of elements in powerRelation$elements that are indifferent to one another.

Details

Similar to PowerRelation(), SocialRanking expects expects a list to represent a power relation. Unlike PowerRelation() however, this list should not be nested and should only contain vectors, each vector containing elements that are deemed equally preferable.

Use doRanking() to rank elements based on arbitrary score objects.

A social ranking solution, or ranking solution, or solution, maps each power relation between coalitions to a power relation between its elements. I.e., from the power relation \(\succsim: \{1,2\} \succ \{2\} \succ \{1\}\), we may expect the result of a ranking solution \(R^\succsim\) to rank element 2 over 1. Therefore \(2 R^\succsim 1\) will be present, but not \(1 R^\succsim 2\).

Formally, a ranking solution \(R: \mathcal{T}(\mathcal{P}) \rightarrow \mathcal{T}(N)\) is a function that, given a power relation \(\succsim \in \mathcal{T}(\mathcal{P})\), always produces a power relation \(R(\succsim)\) (or \(R^\succsim\)) over its set of elements. For two elements \(i, j \in N\), \(i R^\succsim j\) means that applying the solution \(R\) on the ranking \(\succsim\) makes \(i\) at least as preferable as \(j\). Often times \(iI^\succsim j\) and \(iP^\succsim j\) are used to indicate its symmetric and asymmetric part, respectively. As in, \(iI^\succsim j\) implies that \(iR^\succsim j\) and \(jR^\succsim i\), whereas \(iP^\succsim j\) implies that \(iR^\succsim j\) but not \(jR^\succsim i\).

See also

Examples

SocialRanking(list(c("a", "b"), "f", c("c", "d")))
#> a ~ b > f > c ~ d
# a ~ b > f > c ~ d