L1 Scores

Calculate the \(L^{(1)}\) scores.

Usage

L1Scores(powerRelation, elements = powerRelation$elements)

L1Ranking(powerRelation)

lexcel1Scores(powerRelation, elements = powerRelation$elements)

lexcel1Ranking(powerRelation)

Arguments

powerRelation

A PowerRelation object created by PowerRelation() or as.PowerRelation()

elements

Vector of elements of which to calculate their scores. By default, the scores of all elements in powerRelation$elements are considered.

Value

Score function returns a list of type Lexcel1Scores and length of powerRelation$elements

(unless parameter elements is specified). Each index contains a vector of length powerRelation$eqs, the number of times the given element appears in each equivalence class.

Ranking function returns corresponding SocialRankingSolution object.

Details

Similar to lexcelRanking(), the number of times an element appears in each equivalence class is counted. In addition, we now also consider the size of the coalitions.

Let \(N\) be a set of elements, \(\succeq \in \mathcal{T}(\mathcal{P})\) be a power relation, and \(\Sigma_1 \succ \Sigma_2 \succ \dots \succ \Sigma_m\) its corresponding quotient order.

For an element \(i \in N\), we get a matrix \(M^\succeq_i\) with \(m\) columns and \(|N|\) rows. Whereas each column represents an equivalence class, each row corresponds to the coalition size.

$$(M^\succeq_i)_{pq} = |\{S \in \Sigma_q: |S| = p\}|$$

Take as an example \(\succeq: (123 \sim 13 \sim 2) \succ (12 \sim 1 \sim 3) \succ (23 \sim \{\})\). From this, we get the following three matrices:

$$ M^\succeq_1 = \begin{bmatrix} 0 & 1 & 0\\ 1 & 1 & 0\\ 1 & 0 & 0 \end{bmatrix} M^\succeq_2 = \begin{bmatrix} 1 & 0 & 0\\ 0 & 1 & 1\\ 1 & 0 & 0 \end{bmatrix} M^\succeq_3 = \begin{bmatrix} 0 & 1 & 0\\ 1 & 0 & 1\\ 1 & 0 & 0 \end{bmatrix} $$

The \(L^{(1)}\) then ranks the elements, rewarding elements that appear in higher ranking coalitions as well as smaller coalitions. When comparing two matrices for a power relation, if \(M^\succeq_i >_{L^{(1)}} M^\succeq_j\), this suggests that there exists a \(p^0 \in \{1, \dots, |N|\}\) and \(q^0 \in \{1, \dots, m\}\) such that the following holds:

\((M^\succeq_i)_{p^0q^0} > (M^\succeq_j)_{p^0q^0}\)

\((M^\succeq_i)_{pq^0} = (M^\succeq_j)_{pq^0}\) for all \(p < p^0\)

\((M^\succeq_i)_{pq} = (M^\succeq_j)_{pq}\) for all \(q < q^0\) and \(p \in \{1, \dots, |N|\}\)

In the aforementioned example, we have that \(M^\succeq_2 >_{L^{(1)}} M^\succeq_1 >_{L^{(1)}} M^\succeq_3\), suggesting that element 2 should be ranked highest and 3 lowest.

Aliases

For better discoverability, lexcel1Scores() and lexcel1Ranking() serve as aliases for L1Scores() and L1Ranking(), respectively.

References

Algaba E, Moretti S, Rémila E, Solal P (2021). “Lexicographic solutions for coalitional rankings.” Social Choice and Welfare, 57(4), 1--33.

See also

Examples

pr <- as.PowerRelation("(123 ~ 13 ~ 2) > (12 ~ 1 ~ 3) > (23 ~ {})")
scores <- L1Scores(pr)
scores$`1`
#>      [,1] [,2] [,3]
#> [1,]    0    1    0
#> [2,]    1    1    0
#> [3,]    1    0    0
#      [,1] [,2] [,3]
# [1,]    0    1    0
# [2,]    1    1    0
# [3,]    1    0    0

L1Ranking(pr)
#> 2 > 1 > 3
# 2 > 1 > 3